components/calendar_client_service_api_client/pyproject.toml
[project]
name = "calendar-client-service-api-client"
version = "0.1.0"
description = "Auto-generated client for calendar_client_service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.23.0,<0.29.0",
    "attrs>=22.2.0",
    "python-dateutil>=2.8.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["calendar_client_service_api_client"]

[tool.ruff]
line-length = 100
target-version = "py311"
extend = "../../pyproject.toml"

[tool.ruff.lint]
ignore = []
